Refactor data_connector.Request -> model.Record
